+++ src/hooks/useSyncEntityNavigation.ts
